requirements.txt
numpy==1.19.4
torchvision
tqdm==4.56.0
more_itertools==8.7.0
moviepy==1.0.3
ffmpeg_python==0.2.0
torch
scikit_image>=0.17.2
requests==2.25.1
hsh==1.1.0
Pillow==8.1.1
opencv-python-headless==4.5.3.56
